+++ NEWS.md
@@ -1,6 +1,7 @@
 # openair 2.8-99
 
 - fix annotation bug when comparing two data sets in `windRose`
+- enhance `selectRunning`. Now returns full data frame with a new condition column.
 
 # openair 2.8-4